requirements.txt
numpy
matplotlib
torch
torchvision
imageio
pillow
pandas
seaborn
SciPy
scikit-learn
mpi4py
apex